+++ plugin-maven/src/main/java/com/diffplug/spotless/maven/FormatterFactory.java
@@ -1,5 +1,5 @@
 /*
- * Copyright 2016-2021 DiffPlug
+ * Copyright 2016-2022 DiffPlug
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -15,6 +15,7 @@
  */
 package com.diffplug.spotless.maven;
 
+import static com.diffplug.spotless.maven.AbstractSpotlessMojo.RATCHETFROM_NONE;
 import static java.util.Collections.emptySet;
 
 import java.io.File;
@@ -35,7 +36,17 @@ import com.diffplug.spotless.Formatter;
 import com.diffplug.spotless.FormatterStep;
 import com.diffplug.spotless.LineEnding;
 import com.diffplug.spotless.generic.PipeStepPair;
-import com.diffplug.spotless.maven.generic.*;
+import com.diffplug.spotless.maven.generic.EclipseWtp;
+import com.diffplug.spotless.maven.generic.EndWithNewline;
+import com.diffplug.spotless.maven.generic.Indent;
+import com.diffplug.spotless.maven.generic.Jsr223;
+import com.diffplug.spotless.maven.generic.LicenseHeader;
+import com.diffplug.spotless.maven.generic.NativeCmd;
+import com.diffplug.spotless.maven.generic.Prettier;
+import com.diffplug.spotless.maven.generic.Replace;
+import com.diffplug.spotless.maven.generic.ReplaceRegex;
+import com.diffplug.spotless.maven.generic.ToggleOffOn;
+import com.diffplug.spotless.maven.generic.TrimTrailingWhitespace;
 
 public abstract class FormatterFactory {
 	@Parameter
@@ -159,6 +170,8 @@ public abstract class FormatterFactory {
 	Optional<String> ratchetFrom(FormatterConfig config) {
 		if (RATCHETFROM_NOT_SET_AT_FORMAT_LEVEL.equals(ratchetFrom)) {
 			return config.getRatchetFrom();
+		} else if (RATCHETFROM_NONE.equals(ratchetFrom)) {
+			return Optional.empty();
 		} else {
 			return Optional.ofNullable(ratchetFrom);
 		}